# Add parallel coset shift method (speeds up coset_fft), add fft correctness tests